requirements.txt
fastapi
uvicorn
SQLAlchemy
Jinja2
argon2-cffi
python-multipart
neat-python